Get Content Items

Returns content items as a data frame from the Connect server. It will only return content that is visible to the API key's user account.

content(client, unpublished = FALSE)

Arguments

client
A Client object (see connect)

unpublished
A boolean value specifying whether to return content that has not successfully published

Value

A data frame (tibble) of content items

id - Auto-incrementing identifier for each content item (legacy)

guid - Unique identifier for each content item (preferred)

app_mode - The type of the content item (examples: shiny, rmd-static, static, python-dash, etc.)

content_category - For static app modes, the specific category of content (examples: site, plot, pin, etc.)

name - The name of the content item as set at initial publishing

title - The user-provided title of the content item

description - The user-provided description of the content item

url - The URL to the content item

owner_guid - Unique identifier of the owner of the content item

owner_username - Username of the owner of the content item

owner_first_name - First name of the owner of the content item

owner_last_name - Last name of the owner of the content item

tags - A data frame of the tags associated with the content item, with format: (id, name, parent_id, created_time, updated_time)

created_time - The timestamp at which the content item was created

updated_time - The timestamp at which the content item was last updated